storage: implement retention for moderation

+++ pkg/storage/storage.go
@@ -2,6 +2,7 @@ package storage
 
 import (
 	"context"
+	"errors"
 	"fmt"
 	"os"
 	"time"
@@ -13,14 +14,17 @@ import (
 	"stream.place/streamplace/pkg/model"
 )
 
+const moderationRetention = 120 * time.Second
+
 func StartSegmentCleaner(ctx context.Context, mod model.Model, cli *config.CLI) error {
+	ctx = log.WithLogValues(ctx, "func", "StartSegmentCleaner")
 	g, ctx := errgroup.WithContext(ctx)
 	g.Go(func() error {
 		for {
 			select {
 			case <-ctx.Done():
 				return nil
-			case <-time.After(10 * time.Second):
+			case <-time.After(60 * time.Second):
 				expiredSegments, err := mod.GetExpiredSegments(ctx)
 				if err != nil {
 					return err
@@ -44,13 +48,17 @@ func StartSegmentCleaner(ctx context.Context, mod model.Model, cli *config.CLI) 
 }
 
 func deleteSegment(ctx context.Context, mod model.Model, cli *config.CLI, seg model.Segment) error {
+	if time.Since(seg.StartTime) < moderationRetention {
+		log.Debug(ctx, "Skipping deletion of segment", "id", seg.ID, "time since start", time.Since(seg.StartTime))
+		return nil
+	}
 	aqt := aqtime.FromTime(seg.StartTime)
 	fpath, err := cli.SegmentFilePath(seg.RepoDID, fmt.Sprintf("%s.%s", aqt.FileSafeString(), "mp4"))
 	if err != nil {
 		return err
 	}
 	err = os.Remove(fpath)
-	if err != nil {
+	if err != nil && !errors.Is(err, os.ErrNotExist) {
 		return err
 	}
 	err = mod.DeleteSegment(ctx, seg.ID)
